Add click animation on user page
Previously, the user wouldn't get any kind of feedback when using the
page. This could create problems where the user would repeatedly click
until the client replied with a new image.

Now, on every `mousedown` the user fires on the `media-container` div,
a small click animation will play, indicating the user that the event
got registered.